Move project summary files to directory
