+++ moped-editor/src/views/projects/projectView/ProjectSummary/ProjectSummaryMap.js
@@ -15,7 +15,7 @@ import {
   countFeatures,
   useHoverLayer,
   useFeatureCollectionToFitBounds,
-} from "../../../utils/mapHelpers";
+} from "../../../../utils/mapHelpers";
 
 const useStyles = makeStyles({
   locationCountText: {
